apps/automation-api/app/requirements.txt
fastapi
uvicorn[standard]
boto3>=1.26.0
python-dotenv
ansible-core>=2.12.0
ansible-runner
pyyaml>=6.0
jinja2>=3.0.0
paramiko

